+++ src/extension.ts
@@ -1,7 +1,9 @@
 import * as vscode from 'vscode';
 
-import { initCommands } from './commands/init';
+import { initTaskCmd } from './commands/tasks_commands';
 import { taskViewDataProvider} from './view/tasks_view';
+import { initConfigView } from './view/config_view';
+import { initConfigManager } from './utils/config_manager';
 
 export var extensionRootUri: vscode.Uri;
 export var userName: string;
@@ -11,7 +13,9 @@ export function activate(context: vscode.ExtensionContext) {
 
 	extensionRootUri = vscode.Uri.file(context.extensionPath);
 
-	initCommands(context);
+	initConfigManager(context);
+
+	initTaskCmd(context);
 
 	context.subscriptions.push(vscode.window.createTreeView('slurm--_tasks_view', { treeDataProvider: taskViewDataProvider }));
 	// context.subscriptions.push(vscode.window.createTreeView('slurm--_config_view', { treeDataProvider: taskViewDataProvider }));